[lldb/Swift] Separate Swift logging from core logging (NFC)
This patch removes the Swift logging channel from lldb's core logging file.
It moves it into a separate file in the Swift Language Plugin, so it can be
build conditionally when Swift support is enabled.

This also applies to the the `swift-healthcheck` command object, which
will now only be available when Swift support is enabled.
