+++ lldb/source/Plugins/TypeSystem/Swift/SwiftASTContext.cpp
@@ -126,6 +126,7 @@
 
 #include "llvm/ADT/ScopeExit.h"
 
+#include "Plugins/Language/Swift/LogChannelSwift.h"
 #include "Plugins/LanguageRuntime/Swift/SwiftLanguageRuntime.h"
 #include "Plugins/Platform/MacOSX/PlatformDarwin.h"
 #include "Plugins/SymbolFile/DWARF/DWARFASTParserClang.h"
